[Modify] 피드페이지 타입 분리
- feed-card 타입 분리

Resolves #65
