+++ src/features/feed/types/feed-card-props.ts
@@ -1,0 +1,15 @@
+export interface FeedCardProps {
+  id: number;
+  user: {
+    nickname: string;
+    image: string;
+  };
+  post: {
+    content: string;
+    imageUrl: string;
+    date: string;
+  };
+  onExpand?: () => void;
+  onCollapse?: () => void;
+  className?: string;
+}
